fix #6 refactor the code
